backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
sqlparse==0.4.3
django-ninja==0.20.0
whitenoise==6.3.0
django-cors-headers==3.13.0
mypy==1.0.0
django-stubs==1.14.0
